fixup! Call Connect() before StartTls for Linux
