+++ src/libraries/System.DirectoryServices.Protocols/src/System/DirectoryServices/Protocols/ldap/LdapSessionOptions.Linux.cs
@@ -1,7 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Collections;
 using System.ComponentModel;
 using System.Runtime.InteropServices;
 
@@ -54,201 +53,13 @@ namespace System.DirectoryServices.Protocols
             }
         }
 
-        public unsafe void StartTransportLayerSecurity(DirectoryControlCollection controls)
+        private void InternalStartTlsHelper()
         {
-            IntPtr serverControlArray = IntPtr.Zero;
-            LdapControl[] managedServerControls = null;
-            IntPtr clientControlArray = IntPtr.Zero;
-            LdapControl[] managedClientControls = null;
-            IntPtr ldapResult = IntPtr.Zero;
-            IntPtr referral = IntPtr.Zero;
-
-            int serverError = 0;
-            Uri[] responseReferral = null;
-
-            if (_connection._disposed)
-            {
-                throw new ObjectDisposedException(GetType().Name);
-            }
-
             // Connect first to make sure LDAP URI is set up properly.
             if (!_connection._connected)
             {
                 _connection.Connect();
                 _connection._connected = true;
-            }
-
-            try
-            {
-                IntPtr tempPtr = IntPtr.Zero;
-
-                // build server control
-                managedServerControls = _connection.BuildControlArray(controls, true);
-                int structSize = Marshal.SizeOf(typeof(LdapControl));
-                if (managedServerControls != null)
-                {
-                    serverControlArray = Utility.AllocHGlobalIntPtrArray(managedServerControls.Length + 1);
-                    for (int i = 0; i < managedServerControls.Length; i++)
-                    {
-                        IntPtr controlPtr = Marshal.AllocHGlobal(structSize);
-                        Marshal.StructureToPtr(managedServerControls[i], controlPtr, false);
-                        tempPtr = (IntPtr)((long)serverControlArray + IntPtr.Size * i);
-                        Marshal.WriteIntPtr(tempPtr, controlPtr);
-                    }
-
-                    tempPtr = (IntPtr)((long)serverControlArray + IntPtr.Size * managedServerControls.Length);
-                    Marshal.WriteIntPtr(tempPtr, IntPtr.Zero);
-                }
-
-                // Build client control.
-                managedClientControls = _connection.BuildControlArray(controls, false);
-                if (managedClientControls != null)
-                {
-                    clientControlArray = Utility.AllocHGlobalIntPtrArray(managedClientControls.Length + 1);
-                    for (int i = 0; i < managedClientControls.Length; i++)
-                    {
-                        IntPtr controlPtr = Marshal.AllocHGlobal(structSize);
-                        Marshal.StructureToPtr(managedClientControls[i], controlPtr, false);
-                        tempPtr = (IntPtr)((long)clientControlArray + IntPtr.Size * i);
-                        Marshal.WriteIntPtr(tempPtr, controlPtr);
-                    }
-
-                    tempPtr = (IntPtr)((long)clientControlArray + IntPtr.Size * managedClientControls.Length);
-                    Marshal.WriteIntPtr(tempPtr, IntPtr.Zero);
-                }
-
-                int error = LdapPal.StartTls(_connection._ldapHandle, ref serverError, ref ldapResult, serverControlArray, clientControlArray);
-                if (ldapResult != IntPtr.Zero)
-                {
-                    // Parse the referral.
-                    int resultError = LdapPal.ParseResultReferral(_connection._ldapHandle, ldapResult, IntPtr.Zero, IntPtr.Zero, IntPtr.Zero, ref referral, IntPtr.Zero, 0 /* not free it */);
-                    if (resultError == 0 && referral != IntPtr.Zero)
-                    {
-                        char** referralPtr = (char**)referral;
-                        char* singleReferral = referralPtr[0];
-                        int i = 0;
-                        ArrayList referralList = new ArrayList();
-                        while (singleReferral != null)
-                        {
-                            string s = LdapPal.PtrToString((IntPtr)singleReferral);
-                            referralList.Add(s);
-
-                            i++;
-                            singleReferral = referralPtr[i];
-                        }
-
-                        // Free heap memory.
-                        if (referral != IntPtr.Zero)
-                        {
-                            LdapPal.FreeValue(referral);
-                            referral = IntPtr.Zero;
-                        }
-
-                        if (referralList.Count > 0)
-                        {
-                            responseReferral = new Uri[referralList.Count];
-                            for (int j = 0; j < referralList.Count; j++)
-                            {
-                                responseReferral[j] = new Uri((string)referralList[j]);
-                            }
-                        }
-                    }
-                }
-
-                if (error != (int)ResultCode.Success)
-                {
-                    if (Utility.IsResultCode((ResultCode)error))
-                    {
-                        // If the server failed request for whatever reason, the ldap_start_tls returns LDAP_OTHER
-                        // and the ServerReturnValue will contain the error code from the server.
-                        if (error == (int)ResultCode.Other)
-                        {
-                            error = serverError;
-                        }
-
-                        string errorMessage = OperationErrorMappings.MapResultCode(error);
-                        ExtendedResponse response = new ExtendedResponse(null, null, (ResultCode)error, errorMessage, responseReferral);
-                        response.ResponseName = "1.3.6.1.4.1.1466.20037";
-                        throw new TlsOperationException(response);
-                    }
-                    else if (LdapErrorMappings.IsLdapError(error))
-                    {
-                        string errorMessage = LdapErrorMappings.MapResultCode(error);
-                        throw new LdapException(error, errorMessage);
-                    }
-                }
-            }
-            finally
-            {
-                if (serverControlArray != IntPtr.Zero)
-                {
-                    // Release the memory from the heap.
-                    for (int i = 0; i < managedServerControls.Length; i++)
-                    {
-                        IntPtr tempPtr = Marshal.ReadIntPtr(serverControlArray, IntPtr.Size * i);
-                        if (tempPtr != IntPtr.Zero)
-                        {
-                            Marshal.FreeHGlobal(tempPtr);
-                        }
-                    }
-                    Marshal.FreeHGlobal(serverControlArray);
-                }
-
-                if (managedServerControls != null)
-                {
-                    for (int i = 0; i < managedServerControls.Length; i++)
-                    {
-                        if (managedServerControls[i].ldctl_oid != IntPtr.Zero)
-                        {
-                            Marshal.FreeHGlobal(managedServerControls[i].ldctl_oid);
-                        }
-
-                        if (managedServerControls[i].ldctl_value != null)
-                        {
-                            if (managedServerControls[i].ldctl_value.bv_val != IntPtr.Zero)
-                            {
-                                Marshal.FreeHGlobal(managedServerControls[i].ldctl_value.bv_val);
-                            }
-                        }
-                    }
-                }
-
-                if (clientControlArray != IntPtr.Zero)
-                {
-                    // Release the memory from the heap.
-                    for (int i = 0; i < managedClientControls.Length; i++)
-                    {
-                        IntPtr tempPtr = Marshal.ReadIntPtr(clientControlArray, IntPtr.Size * i);
-                        if (tempPtr != IntPtr.Zero)
-                        {
-                            Marshal.FreeHGlobal(tempPtr);
-                        }
-                    }
-
-                    Marshal.FreeHGlobal(clientControlArray);
-                }
-
-                if (managedClientControls != null)
-                {
-                    for (int i = 0; i < managedClientControls.Length; i++)
-                    {
-                        if (managedClientControls[i].ldctl_oid != IntPtr.Zero)
-                        {
-                            Marshal.FreeHGlobal(managedClientControls[i].ldctl_oid);
-                        }
-
-                        if (managedClientControls[i].ldctl_value != null)
-                        {
-                            if (managedClientControls[i].ldctl_value.bv_val != IntPtr.Zero)
-                                Marshal.FreeHGlobal(managedClientControls[i].ldctl_value.bv_val);
-                        }
-                    }
-                }
-
-                if (referral != IntPtr.Zero)
-                {
-                    LdapPal.FreeValue(referral);
-                }
             }
         }
